Move abilities to isolated component
