+++ src/components/About.js
@@ -1,10 +1,10 @@
-/* eslint-disable no-param-reassign */
 import { Box } from '@mui/material';
-import React, { useEffect } from 'react';
+import React from 'react';
 import { useSelector } from 'react-redux';
 import { v4 as uuidv4 } from 'uuid';
 import Skill from './Skill';
 import animation from '../assets/gifs/learning.gif';
+import Ability from './Ability';
 
 const About = () => {
   const { about } = useSelector((state) => state);
@@ -18,38 +18,6 @@ const About = () => {
     reading,
     writing,
   } = about;
-
-  useEffect(() => {
-    const learningContainer = document.querySelector('.learning');
-    const elements = document.querySelectorAll('.ability');
-
-    elements.forEach((element) => {
-      const width = learningContainer.offsetWidth - element.offsetWidth;
-
-      const rect = element.getBoundingClientRect();
-      let position = rect.left;
-      let pace = 1;
-
-      const move = () => {
-        window.requestAnimationFrame(move);
-        position += pace;
-
-        if (position >= width) {
-          pace = -1;
-          position += pace;
-        }
-
-        if (position <= 0) {
-          pace = 1;
-          position += pace;
-        }
-
-        element.style.left = `${position}px`;
-      };
-
-      move();
-    });
-  }, []);
 
   return (
     <Box className="flex flex-col gap-10 items-center text-primary-1000 pt-5 px-5">
@@ -79,12 +47,7 @@ const About = () => {
         <ul className="flex flex-col gap-3 learning w-full">
           {
             learning.map((thing) => (
-              <li
-                key={uuidv4()}
-                className="relative text-primary-500 left-5 text-lg font-medium ability w-fit"
-              >
-                {thing}
-              </li>
+              <Ability key={uuidv4()} skill={thing} />
             ))
           }
         </ul>
